# perf: add an environment variable to be able to configure a data type of slots and nodes mapping