## Iteration per second
### VS
```
################################################################################
# Mappings between slots and nodes
################################################################################

Warming up --------------------------------------
           CharArray   783.598k i/100ms
               Array     1.849M i/100ms
Calculating -------------------------------------
           CharArray      7.804M (± 2.4%) i/s -     39.180M in   5.024004s
               Array     18.948M (± 3.8%) i/s -     96.166M in   5.083967s

Comparison:
               Array: 18947756.4 i/s
           CharArray:  7803935.6 i/s - 2.43x  slower
```

### Array
```
################################################################################
# pipelined
################################################################################

Warming up --------------------------------------
 pipelined: ondemand    50.000  i/100ms
   pipelined: pooled    53.000  i/100ms
     pipelined: none    53.000  i/100ms
    pipelined: envoy    55.000  i/100ms
   pipelined: cproxy    47.000  i/100ms
Calculating -------------------------------------
 pipelined: ondemand    483.238  (±10.1%) i/s -      2.400k in   5.030776s
   pipelined: pooled    529.960  (± 8.9%) i/s -      2.650k in   5.054174s
     pipelined: none    519.811  (± 6.7%) i/s -      2.650k in   5.125332s
    pipelined: envoy    553.978  (± 6.9%) i/s -      2.805k in   5.095757s
   pipelined: cproxy    366.108  (±21.0%) i/s -      1.786k in   5.136825s

Comparison:
    pipelined: envoy:      554.0 i/s
   pipelined: pooled:      530.0 i/s - same-ish: difference falls within error
     pipelined: none:      519.8 i/s - same-ish: difference falls within error
 pipelined: ondemand:      483.2 i/s - same-ish: difference falls within error
   pipelined: cproxy:      366.1 i/s - 1.51x  slower
```

### CharArray
```
################################################################################
# pipelined
################################################################################

Warming up --------------------------------------
 pipelined: ondemand    41.000  i/100ms
   pipelined: pooled    60.000  i/100ms
     pipelined: none    61.000  i/100ms
    pipelined: envoy    63.000  i/100ms
   pipelined: cproxy    40.000  i/100ms
Calculating -------------------------------------
 pipelined: ondemand    561.485  (± 8.9%) i/s -      2.788k in   5.020871s
   pipelined: pooled    608.369  (± 8.1%) i/s -      3.060k in   5.077427s
     pipelined: none    622.878  (± 1.6%) i/s -      3.172k in   5.093801s
    pipelined: envoy    628.646  (± 9.5%) i/s -      3.150k in   5.074765s
   pipelined: cproxy    490.748  (±13.7%) i/s -      2.400k in   5.047291s

Comparison:
    pipelined: envoy:      628.6 i/s
     pipelined: none:      622.9 i/s - same-ish: difference falls within error
   pipelined: pooled:      608.4 i/s - same-ish: difference falls within error
 pipelined: ondemand:      561.5 i/s - same-ish: difference falls within error
   pipelined: cproxy:      490.7 i/s - 1.28x  slower
```

It seems that the bottleneck is another location. Envoy tops us in pipelining.

## Memory profiling
### Array
```
################################################################################
# pipelining_in_moderation: primary_only
################################################################################

Total allocated: 2711216 bytes (45335 objects)
Total retained:  120 bytes (3 objects)

allocated memory by gem
-----------------------------------
   1867194  redis-client-0.17.0
    563283  redis-cluster-client/lib
    172280  other
     96987  socket
     11472  uri

allocated memory by file
-----------------------------------
    799044  redis-client-0.17.0/lib/redis_client/ruby_connection/resp3.rb
    760182  redis-client-0.17.0/lib/redis_client/ruby_connection/buffered_io.rb
    313579  redis-cluster-client/lib/redis_client/cluster/node.rb
    280160  redis-client-0.17.0/lib/redis_client/command_builder.rb
    169640  /home/runner/work/redis-cluster-client/redis-cluster-client/test/prof_mem.rb
```

### CharArray
```
################################################################################
# pipelining_in_moderation: primary_only
################################################################################

Total allocated: 2462301 bytes (45341 objects)
Total retained:  120 bytes (3 objects)

allocated memory by gem
-----------------------------------
   1867194  redis-client-0.17.0
    335053  redis-cluster-client/lib
    172280  other
     76302  socket
     11472  uri

allocated memory by file
-----------------------------------
    799044  redis-client-0.17.0/lib/redis_client/ruby_connection/resp3.rb
    760182  redis-client-0.17.0/lib/redis_client/ruby_connection/buffered_io.rb
    280160  redis-client-0.17.0/lib/redis_client/command_builder.rb
    169640  /home/runner/work/redis-cluster-client/redis-cluster-client/test/prof_mem.rb
     85541  redis-cluster-client/lib/redis_client/cluster/node.rb
```